requirements.txt
google-generativeai
streamlit
Pillow
python-dateutil
python-docx
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
apscheduler
firebase-admin
requests
pydantic
